backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
asyncpg==0.30.0
pgvector==0.3.6
azure-storage-blob==12.24.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
onnxruntime==1.20.1
insightface==0.7.3
